fmt: apply rustfmt to source and test modules, remove stray #[test] attributes
Mechanical formatting pass across ~150 source and test files:
- Break long assert!/assert_eq!/assert!(matches!(...)) calls across lines
- Wrap long function arguments and struct construction to line-width limits
- Re-sort use imports alphabetically (Rust 2024 edition style)
- Expand single-line struct literals to multi-line where needed
- Remove 21 orphaned #[test] attributes sitting above comment blocks
  rather than above actual test functions (bulkhead, gc, checkpoint, etc.)

No semantic changes.

+++ crates/franken-engine/src/portfolio_governor/governance_audit_ledger.rs
@@ -1711,7 +1711,7 @@ mod tests {
                     i * 10,
                 ))
                 .unwrap();
-            assert_eq!(entry.sequence, i as u64);
+            assert_eq!(entry.sequence, i);
         }
     }
 }
